Day 7 part 2

+++ 2022/7/Classes.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace _7;
 
-public record File(string Name, long Size);
+public record File(string Name, int Size);
 public record Directory(string Name, List<File> Files, List<Directory> SubDirectories, IDirectory Parent) : IDirectory;
 public record RootDirectory(List<File> Files, List<Directory> SubDirectories) : IDirectory;
